[fix] instance, auth api 수정 (#242)
