+++ app/_common/api/auth.ts
@@ -49,8 +49,6 @@ export const getSignUpUser = async () => {
 
 export const reIssueAccessToken = async (refreshToken: string) => {
   try {
-    if (refreshToken === "") return navigate("/login");
-
     const { data } = await instance.post<reIssueAccessTokenResponse>(
       "/auth/reissue",
       {
@@ -66,6 +64,7 @@ export const reIssueAccessToken = async (refreshToken: string) => {
 
     toast.error("인증이 만료되었습니다. 재로그인이 필요합니다.");
 
-    navigate("/login");
+    // TODO: 백엔드 reissue api 수정 후 주석 해제
+    // navigate("/login");
   }
 };
