+++ app/_common/api/instance.ts
@@ -1,7 +1,5 @@
 import axios from "axios";
 
-import { getCookie } from "../utils/cookie";
-import { reIssueAccessToken } from "./auth";
 const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
 const SERVER_VERSION = "/api/v1";
 
@@ -28,25 +26,24 @@ instance.interceptors.response.use(
     return response;
   },
   async error => {
-    const { status, config } = error.response;
-    // TODO: 응답 헤더의 Set-Cookie로 refreshToken 저장으로 변경되어 삭제 필요?
-    const refreshToken = await getCookie("refreshToken", "");
+    // TODO: refreshtoken, cookie 버그 해결될 때까지 주석 유지
+    // const { status, config } = error.response;
 
-    if (status === 401) {
-      if (config.url !== "/auth/reissue") {
-        const newAccessToken = await reIssueAccessToken(refreshToken);
+    // const refreshToken = await getCookie("refreshToken", "");
 
-        if (newAccessToken) {
-          localStorage.setItem("accessToken", newAccessToken);
+    // if (status === 404) {
+    //   const newAccessToken = await reIssueAccessToken(refreshToken);
 
-          config.headers.Authorization = `Bearer ${newAccessToken}`;
+    //   if (newAccessToken) {
+    //     localStorage.setItem("accessToken", newAccessToken);
 
-          return instance(config);
-        }
-      }
+    //     config.headers.Authorization = `Bearer ${newAccessToken}`;
 
-      return Promise.reject(error);
-    }
+    //     return instance(config);
+    //   }
+
+    //   return Promise.reject(error);
+    // }
 
     return Promise.reject(error);
   },
