+++ app/login/_hooks/useMutationLogin.ts
@@ -3,20 +3,39 @@ import { useMutation } from "@tanstack/react-query";
 
 import { postLogin } from "@/app/_common/api/auth";
 
-import { reIssueAccessTokenResponse } from "../_types/login.types";
+import { setCookie } from "@/app/_common/utils/cookie";
+
+import { LoginResponse } from "../_types/login.types";
+
+const saveNewTokens = (newAccessToken: string, newRefreshToken: string) => {
+  sessionStorage.setItem("accessToken", newAccessToken);
+  sessionStorage.setItem("refreshToken", newRefreshToken);
+};
+
+const saveUpdatedTokens = (newAccessToken: string, newRefreshToken: string) => {
+  localStorage.setItem("accessToken", newAccessToken);
+
+  setCookie("refreshToken", newRefreshToken, {
+    httpOnly: true,
+    maxAge: 60 * 60 * 24 * 14,
+  });
+};
 
 export const useMutationLogin = () => {
   const router = useRouter();
   const mutation = useMutation({
     mutationFn: (code: string) => postLogin(code),
-    onSuccess: (data: reIssueAccessTokenResponse) => {
+    onSuccess: (data: LoginResponse) => {
       if (data.signUpComplete) {
-        localStorage.setItem("accessToken", data.accessToken);
+        saveUpdatedTokens(data.accessToken, data.refreshToken);
         router.push("/");
       } else {
-        sessionStorage.setItem("accessToken", data.accessToken);
+        saveNewTokens(data.accessToken, data.refreshToken);
         router.push("/signup");
       }
+    },
+    onError: () => {
+      console.error("로그인 에러 발생");
     },
   });
